fix(geometry): re-registering the skip-dependent processors no longer discards their logs (#3821)
`set_skip_small_cuts` re-registers the boolean and CSG processors so they pick
up the new flag. `register` overwrites the map entries, which DROPS the old
processors — and a dropped processor takes its failure log with it. The viewer
flips this flag on a live router, so any record made before the flip and not yet
drained was lost silently.

`register_skip_dependent_processors` now sweeps first. A drain rather than a
`debug_assert`: the assert would only catch the loss in a debug build, and the
records are recoverable, so recovering them is strictly better than detecting
their disposal.

RED/GREEN: `flipping_skip_small_cuts_does_not_discard_an_undrained_log` records
a failure, flips the flag WITHOUT draining, then drains; removing the sweep
fails it and nothing else.

Also moved `UNATTRIBUTED_PRODUCT_ID` and `count_attributed_products` to
diagnostics/attribution.rs so router/diagnostics.rs stays under budget
(586 -> 575), and refreshed the boolean/mod.rs and router/mod.rs rows to their
real sizes. Every row still sits below what it was on main; the rust/geometry
digest moves with them.
